+++ kamu-core/src/infra/utils/docker_images.rs
@@ -1,6 +1,6 @@
-pub const SPARK: &str = "kamudata/engine-spark:0.10.1";
+pub const SPARK: &str = "kamudata/engine-spark:0.10.2";
 pub const LIVY: &str = SPARK;
-pub const FLINK: &str = "kamudata/engine-flink:0.7.0";
+pub const FLINK: &str = "kamudata/engine-flink:0.8.0-flink_1.13.1-scala_2.12-java8";
 pub const JUPYTER: &str = "kamudata/jupyter-uber:0.0.1";
 
 // Test Images