+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.42.0] - 2021-06-05
+### Changed
+- Upgraded Flink engine to latest `v1.13.1`
+- Improved empty data block handling in metadata
+- Improved some CLI output format
+
 ## [0.41.0] - 2021-05-08
 ### Added
 - `zsh` completions support (via `bashcompinit`)